+++ src/main/java/com/poc/ecard/controller/UserOps.java
@@ -1,24 +1,25 @@
 package com.poc.ecard.controller;
 
 
+import com.google.zxing.WriterException;
+import com.poc.ecard.dtos.User;
+import com.poc.ecard.entity.UserDetails;
 import com.poc.ecard.dtos.ContactsBookReq;
 import com.poc.ecard.dtos.ContactsBookResponse;
-import com.poc.ecard.dtos.FindCommonContactsReq;
-import com.poc.ecard.dtos.FindCommonContactsResponse;
-import com.poc.ecard.service.UserProfileServicesImpl;
+import com.poc.ecard.service.UserServices;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import java.io.IOException;
+import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1")
 public class UserOps {
 
     @Autowired
-    private UserProfileServicesImpl userProfileServices;
+    private UserServices userServices;
 
 
     @PostMapping("/upload_contacts")
@@ -29,11 +30,32 @@ public class UserOps {
     }
 
 
+//    @PostMapping("/find_common_contacts")
+////    @PreAuthorize("hasAnyAuthority('ADMIN','DISTRIBUTOR_ADMIN')")
+//    public ResponseEntity<FindCommonContactsResponse> addUser(@RequestBody FindCommonContactsReq findCommonContactsReq) {
+////        FindCommonContactsResponse response = userServices.findCommonContacts(findCommonContactsReq);
+//        return ResponseEntity.ok(null);
+//    }
+
+    @GetMapping
+    public ResponseEntity<List<UserDetails>> getUsers() throws IOException, WriterException {
+        List<UserDetails> userDetails = userServices.getUser();
+        return ResponseEntity.ok(userServices.getUser());
+    }
+
+    //Mapping to addUser
+    @PostMapping("/add")
+    public ResponseEntity<User> addCustomer(@RequestBody User user) throws IOException, WriterException {
+        userServices.generateByteCode(user);
+        userServices.generateQRCode(user);
+        return ResponseEntity.ok(userServices.addUser(user));
+    }
+
+    //Mapping to get
     @PostMapping("/find_common_contacts")
-//    @PreAuthorize("hasAnyAuthority('ADMIN','DISTRIBUTOR_ADMIN')")
-    public ResponseEntity<FindCommonContactsResponse> addUser(@RequestBody FindCommonContactsReq findCommonContactsReq) {
-//        FindCommonContactsResponse response = userServices.findCommonContacts(findCommonContactsReq);
-        return ResponseEntity.ok(null);
+    public ResponseEntity<User> find_Common_Contacts(@RequestBody User user)
+    {
+        return ResponseEntity.ok(userServices.findCommonContacts(user));
     }
 
 }
