+++ src/main/java/com/poc/ecard/service/UserProfileServicesImpl.java
@@ -1,6 +1,9 @@
 package com.poc.ecard.service;
 
 
+import com.google.zxing.WriterException;
+import com.poc.ecard.dtos.User;
+import com.poc.ecard.entity.UserDetails;
 import com.poc.ecard.dtos.ContactsBookReq;
 import com.poc.ecard.dtos.ContactsBookResponse;
 import com.poc.ecard.dtos.FindCommonContactsReq;
@@ -9,6 +12,9 @@ import com.poc.ecard.repository.UserProfileRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.io.IOException;
+import java.util.List;
 
 @Service
 @Slf4j
@@ -26,6 +32,31 @@ public class UserProfileServicesImpl implements UserServices{
     @Override
     public FindCommonContactsResponse findCommonContacts(FindCommonContactsReq findCommonContactsReq) {
 
+        return null;
+    }
+
+    @Override
+    public List<UserDetails> getUser() {
+        return null;
+    }
+
+    @Override
+    public User addUser(User user) {
+        return null;
+    }
+
+    @Override
+    public User findCommonContacts(User user) {
+        return null;
+    }
+
+    @Override
+    public User generateQRCode(User user) throws WriterException, IOException {
+        return null;
+    }
+
+    @Override
+    public User generateByteCode(User user) throws WriterException, IOException {
         return null;
     }
